requirements.txt
streamlit
pandas
plotly
matplotlib
mysql-connector-python
seaborn